fix: Optimize the execution logic of dynamic forms
