+++ ui/src/components/mk-dynamics-form/constructor/index.vue
@@ -1,193 +1,42 @@
 <template>
-  <template v-if="enableVisibility">
-    <el-tabs v-model="activeTab">
-      <el-tab-pane label="基本信息" name="basic">
-        <el-form
-          @submit.prevent
-          ref="ruleFormRef"
-          class="mb-24"
-          label-width="auto"
-          :model="form_data"
-          v-bind="$attrs"
-        >
-          <el-form-item
-            label="参数"
-            :required="true"
-            prop="field"
-            :rules="rules.field"
-          >
-            <el-input
-              v-model="form_data.field"
-              :maxlength="64"
-              placeholder="请输入参数"
-              show-word-limit
-            />
-          </el-form-item>
-          <el-form-item
-            label="显示名称"
-            :required="true"
-            prop="label"
-            :rules="rules.label"
-          >
-            <el-input
-              v-model="form_data.label"
-              :maxlength="64"
-              show-word-limit
-              placeholder="请输入显示名称"
-            />
-          </el-form-item>
-          <el-form-item label="参数提示说明">
-            <el-input
-              v-model="form_data.tooltip"
-              :maxlength="128"
-              show-word-limit
-              placeholder="请输入参数提示说明"
-            />
-          </el-form-item>
-          <el-form-item
-            label="是否必填"
-            :required="true"
-            prop="required"
-            :rules="rules.required"
-            @click.prevent
-          >
-            <el-switch v-model="form_data.required" :active-value="true" :inactive-value="false" />
-          </el-form-item>
-          <el-form-item
-            label="组件类型"
-            :required="true"
-            prop="input_type"
-            :rules="rules.input_type"
-          >
-            <el-select
-              v-model="form_data.input_type"
-              placeholder="请选择组件类型"
-            >
-              <el-option
-                v-for="input_type in input_type_list"
-                :key="input_type.value"
-                :label="input_type.label"
-                :value="input_type.value"
-              />
-            </el-select>
-          </el-form-item>
-          <component
-            v-if="form_data.input_type && currentConstructor"
-            ref="componentFormRef"
-            v-model="form_data"
-            :is="currentConstructor"
-          ></component>
-        </el-form>
-      </el-tab-pane>
+  <el-tabs v-if="enableVisibility" v-model="activeTab">
+    <el-tab-pane label="基本信息" name="basic">
+      <BasicInfoConstructor
+        ref="basicRef"
+        v-model="form_data"
+        :input-type-list="input_type_list"
+        v-bind="$attrs"
+      />
+    </el-tab-pane>
+    <el-tab-pane label="显隐设置" name="visibility">
+      <VisibilityConstructor
+        ref="visibilityRef"
+        :initialValue="visibility_rules"
+        :leftOptions="leftOptions"
+      />
+    </el-tab-pane>
+  </el-tabs>
 
-      <el-tab-pane label="显隐设置" name="visibility">
-        <VisibilityConstructor
-          ref="visibilityRef"
-          :initialValue="visibility_rules"
-          :leftOptions="leftOptions"
-        />
-      </el-tab-pane>
-    </el-tabs>
-  </template>
-
-  <el-form
+  <BasicInfoConstructor
     v-else
-    @submit.prevent
-    ref="ruleFormRef"
-    class="mb-24"
-    label-width="auto"
-    :model="form_data"
+    ref="basicRef"
+    v-model="form_data"
+    :input-type-list="input_type_list"
     v-bind="$attrs"
-  >
-    <el-form-item
-      label="参数"
-      :required="true"
-      prop="field"
-      :rules="rules.field"
-    >
-      <el-input
-        v-model="form_data.field"
-        :maxlength="64"
-        placeholder="请输入参数"
-        show-word-limit
-      />
-    </el-form-item>
-    <el-form-item
-      label="显示名称"
-      :required="true"
-      prop="label"
-      :rules="rules.label"
-    >
-      <el-input
-        v-model="form_data.label"
-        :maxlength="64"
-        show-word-limit
-        placeholder="请输入显示名称"
-      />
-    </el-form-item>
-    <el-form-item label="参数提示说明">
-      <el-input
-        v-model="form_data.tooltip"
-        :maxlength="128"
-        show-word-limit
-        placeholder="请输入参数提示说明"
-      />
-    </el-form-item>
-    <el-form-item
-      label="是否必填"
-      :required="true"
-      prop="required"
-      :rules="rules.required"
-      @click.prevent
-    >
-      <el-switch v-model="form_data.required" :active-value="true" :inactive-value="false" />
-    </el-form-item>
-    <el-form-item
-      label="组件类型"
-      :required="true"
-      prop="input_type"
-      :rules="rules.input_type"
-    >
-      <el-select
-        v-model="form_data.input_type"
-        placeholder="请选择组件类型"
-      >
-        <el-option
-          v-for="input_type in input_type_list"
-          :key="input_type.value"
-          :label="input_type.label"
-          :value="input_type.value"
-        />
-      </el-select>
-    </el-form-item>
-    <component
-      v-if="form_data.input_type && currentConstructor"
-      ref="componentFormRef"
-      v-model="form_data"
-      :is="currentConstructor"
-    ></component>
-  </el-form>
+  />
 </template>
 
 <script setup lang="ts">
-import { onMounted, ref, nextTick, computed } from 'vue'
-import type { FormInstance } from 'element-plus'
-import _ from 'lodash'
+import { onMounted, ref, nextTick } from 'vue'
 import { input_type_list as input_type_list_data } from '@/components/mk-dynamics-form/constructor/data'
-import VisibilityConstructor from '../visibility/Constructor.vue'
+import VisibilityConstructor from './visibility/index.vue'
+import BasicInfoConstructor from './BasicInfoConstructor.vue'
 import { type LeftOptions } from './type.ts'
 
-// 批量注册构造器组件
-const constructorModules = import.meta.glob('./items/*.vue', { eager: true })
-const constructorComponents: Record<string, any> = {}
-for (const [path, module] of Object.entries(constructorModules)) {
-  const name = path.split('/').pop()?.replace('.vue', '') || ''
-  constructorComponents[name] = (module as any).default
-}
-
-const currentConstructor = computed(() => {
-  return constructorComponents[form_data.value.input_type] || null
-})
+// $attrs（label-position 等）显式透传给 BasicInfoConstructor 的 el-form
+defineOptions({ inheritAttrs: false })
+// 声明 v-model 事件，避免其监听器漏入 $attrs 透传到子表单
+defineEmits(['update:modelValue'])
 
 const props = withDefaults(
   defineProps<{
@@ -205,12 +54,12 @@ const props = withDefaults(
       })),
   },
 )
-const emit = defineEmits(['update:modelValue'])
 
 const activeTab = ref('basic')
-const ruleFormRef = ref<FormInstance>()
+const basicRef = ref<any>()
+const visibilityRef = ref()
+const visibility_rules = ref<any>(null)
 
-const componentFormRef = ref<any>()
 const form_data = ref<any>({
   label: '',
   field: '',
@@ -218,39 +67,16 @@ const form_data = ref<any>({
   required: false,
   input_type: '',
 })
-const visibility_rules = ref<any>(null)
-const visibilityRef = ref()
-const rules = {
-  label: [{ required: true, message: '显示名称 为必填属性' }],
-  field: [{ required: true, message: '参数 为必填属性' }],
-  required: [{ required: true, message: '是否必填 为必填属性' }],
-  input_type: [{ required: true, message: '组建类型 为必填属性' }],
-}
-const getData = () => {
-  let label: string | any = form_data.value.label
-  if (form_data.value.tooltip) {
-    label = {
-      input_type: 'TooltipLabel',
-      label: form_data.value.label,
-      attrs: { tooltip: form_data.value.tooltip },
-      props_info: {},
-    }
-  }
-  return {
-    label: label,
-    required: form_data.value.required,
-    field: form_data.value.field,
-    default_value: form_data.value.default_value,
-    show_default_value: form_data.value.show_default_value,
-    visibility_rules: visibilityRef.value?.getData() ?? null,
-    ...componentFormRef.value.getData(),
-  }
-}
+
+const getData = () => ({
+  ...basicRef.value.getData(),
+  visibility_rules: visibilityRef.value?.getData() ?? null,
+})
 
 const validate = () => {
   const promises = []
-  if (ruleFormRef.value) {
-    promises.push(ruleFormRef.value.validate())
+  if (basicRef.value?.validate) {
+    promises.push(basicRef.value.validate())
   }
   if (visibilityRef.value?.validate) {
     promises.push(visibilityRef.value.validate())
@@ -263,26 +89,12 @@ onMounted(() => {
     rander(props.modelValue)
   }
 })
-const rander = (data: any) => {
-  form_data.value.required = data.required ? data.required : false
-  form_data.value.field = data.field
-  visibility_rules.value = data.visibility_rules ?? null
-  if (data.show_default_value !== undefined) {
-    form_data.value.show_default_value = data.show_default_value
-  }
-  if (data.input_type) {
-    form_data.value.input_type = data.input_type + 'Constructor'
-  }
 
-  if (data.label && data.label.input_type === 'TooltipLabel') {
-    form_data.value.tooltip = data.label.attrs.tooltip
-    form_data.value.label = data.label.label
-  } else {
-    form_data.value.label = data.label
-  }
+const rander = (data: any) => {
+  visibility_rules.value = data.visibility_rules ?? null
   nextTick(() => {
-    componentFormRef.value?.rander(data)
-    visibilityRef.value?.restore(data.visibility_rules)
+    basicRef.value?.rander(data)
+    visibilityRef.value?.rander(data.visibility_rules)
   })
 }
